fix(twitch): oauth token, stream title, title change, upgrade twitch lib
A while ago twitch changed their GET /streams route and it now requires
an oauth token. This change adds a few more config options to support
this. If a client secret is given and an access token is not, an oauth
token will be fetched automatically. The token is refreshed each time
the module is refreshed to ensure an expired token is never used. This
was a simpler solution than tracking the expiration date given in the
oauth token.

This also changes the title of the module. "Top Streams" was kind of
ambiguous.

Upgrading the library used to access the twitch API to v2.